+++ Runtime/ExtensionMethods/ExtensionMethods.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Collections.Generic;
-using System.Linq; 
+using System.Linq;
 using UnityEngine;
 using UnityEngine.Events;
 using Object = UnityEngine.Object;
@@ -1304,9 +1304,6 @@ namespace SOSXR.SeaShark
 
             return persistentCallers;
         }
-
-
-      
 
 
         /// <summary>