+++ Runtime/GeneralAttributes/NoFoldOutAttribute.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 namespace SOSXR.SeaShark.Attributes
 {
     /// <summary>
-    /// From Unity Timeline Samples
-    /// Removes the fold out and draws all child properties inline.
+    ///     From Unity Timeline Samples
+    ///     Removes the fold out and draws all child properties inline.
     /// </summary>
     public class NoFoldOutAttribute : PropertyAttribute
     {